Use KeyName type for KeyName parameter
AWS a while back added this type to parameters. I think it will result in an earlier error if its set to something invalid, and if you create the stack via the aws console, it gives you a dropdown of existing keys rather than a free text field so it is harder to screw up.